WIP - move invoker generator to universalNativeInvoker_x86_64.cpp
