+++ src/hotspot/share/runtime/sharedRuntime.hpp
@@ -510,15 +510,6 @@ class SharedRuntime: AllStatic {
 
   static address handle_unsafe_access(JavaThread* thread, address next_pc);
 
-#ifdef COMPILER2
-  static RuntimeStub* make_native_invoker(address call_target,
-                                          BasicType*,
-                                          int num_args,
-                                          int shadow_space_bytes,
-                                          const GrowableArray<VMReg>& input_registers,
-                                          const GrowableArray<VMReg>& output_registers);
-#endif
-
   static void compute_move_order(const BasicType* in_sig_bt,
                                  int total_in_args, const VMRegPair* in_regs,
                                  int total_out_args, VMRegPair* out_regs,
